Fixed ranged enemy targeting
Swapped the target back to player from shooter. Also added tags for icewall and earthwall for potential future work. Moved ghost into room for easier access for testing. Changed the temp sprite for enemyarrow to something more arrow like and added a check that destroys it if it hits an ice wall or earth wall.